test: extra test for mismatch in python versions
