+++ tests/integration_python/test_edge_cases.py
@@ -1,0 +1,38 @@
+from pathlib import Path
+import pytest
+import platform
+from .common import verify_cli_command, ExitCode
+
+
+@pytest.mark.slow
+def test_pypi_git_deps(pixi: Path, tmp_path: Path) -> None:
+    """Test where we need to lookup recursive git dependencies and consider them first party"""
+    test_data = Path(__file__).parent.parent / "data/pixi_tomls/pip_git_dep.toml"
+    manifest = tmp_path.joinpath("pixi.toml")
+    toml = test_data.read_text()
+    manifest.write_text(toml)
+
+    # Run the installation
+    verify_cli_command(
+        [pixi, "install", "--manifest-path", manifest],
+        ExitCode.SUCCESS,
+    )
+
+
+@pytest.mark.slow
+@pytest.mark.skipif(
+    not (platform.system() == "Darwin" and platform.machine() == "arm64"),
+    reason="Test tailored for macOS arm so that we can get two different python interpreters",
+)
+def test_python_mismatch(pixi: Path, tmp_path: Path) -> None:
+    """Test pypi wheel install where the base interpreter is not the same as the target version"""
+    test_data = Path(__file__).parent.parent / "data/pixi_tomls/python_mismatch.toml"
+    manifest = tmp_path.joinpath("pixi.toml")
+    toml = test_data.read_text()
+    manifest.write_text(toml)
+
+    # Install
+    verify_cli_command(
+        [pixi, "install", "--manifest-path", manifest],
+        ExitCode.SUCCESS,
+    )
